Not permitted by token and probably a bad pattern
If runner can push lockfile to origin then it can push other changes as well.  Meanwhile, the dependency lockfile only changes when the plan changes, meaning it should be subject to merge review. So instead, creator should run terraform init when changes are made and commit the dependency lockfile along with the rest of her code.